Correction after Code Review
