+++ Project X/ViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  ViewModel.swift
+//  Project X
+//
+//  Created by Dmitriy Eni on 27.09.2022.
+//
+
+import UIKit
+
+struct CustomLoginButtonViewModel {
+    let textLabel: String
+    let imageView: String
+    let backgroundColor: UIColor
+    let tintColor: UIColor
+}
